Bloody Autoboxing. This could never have worked; the mapping is of Shorts, not Integers. I guess we don't rely on the SpotScene support.

+++ src/java/com/threerings/msoy/room/data/MsoySceneModel.java
@@ -177,7 +177,7 @@ public class MsoySceneModel extends SceneModel
     public Portal getPortal (int portalId)
     {
         validatePortalInfo();
-        return _portalInfo.get(portalId);
+        return _portalInfo.get(Short.valueOf((short) portalId));
     }
 
     /**
